--- src/main.cpp
@@ -1,6 +1,0 @@
-#include "export.h"
-
-int main()
-{
-	return 0;
-}
