+++ tests/main.cpp
@@ -1,0 +1,6 @@
+#include <image_upscaler/image_upscaler.hpp>
+
+int main()
+{
+	return 0;
+}
